modules/native_libs_module/requirements.txt
redis
androguard>=3.3.5
lief
python-magic